Refactoring. Move abstract base class RepositoryTest to the `shared.japrepo` package.
